Added slides functionality
